# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Built by team C",
    "version": "1.0.0",
    "title": "Desa Sehat",
    "contact": {
      "email": "romadhonatatang@gmail.com"
    }
  },
  "host": "",
  "basePath": "/api",
...
  "paths": {
...
    "/campaign/": {
      "post": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "create new campaign",
        "description": "Ask user for campaign's details in order to make a campaign",
        "operationId": "createCampaign",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "description": "ask user for the picture of the patient",
            "required": true,
            "type": "file"
          },
          {
            "name": "address",
            "in": "formData",
            "description": "ask user for address of the patient",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "ask user for the title of the campaign",
            "required": true,
            "type": "string"
          },
          {
            "name": "patient_name",
            "in": "formData",
            "description": "ask user for the name of the patient",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "ask user for the description of the campaign",
            "required": true,
            "type": "string"
          },
          {
            "name": "total_fund",
            "in": "formData",
            "description": "ask user for how much the fund needed",
            "required": true,
            "type": "number"
          },
          {
            "name": "disease_category",
            "in": "formData",
            "description": "ask user for the category of disease",
            "enum": [
              "Asthma",
              "Diarrhea",
              "Diabetes Mellitus",
              "Dengue Fever",
              "Poor Nutrition",
              "Hypertension",
              "Coronary Heart Disease",
              "Malaria",
              "Stroke",
              "Tuberculosis",
              "Covid19",
              "Dental Health Problem"
            ],
            "required": true,
            "type": "string"
          },
          {
            "name": "due_date",
            "in": "formData",
            "description": "ask user for the deadline of the campaign",
            "required": true,
            "type": "date"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "ask user for the gender of the patient",
            "enum": [
              "Male",
              "Female"
            ],
            "required": true,
            "type": "string"
          },
          {
            "name": "blood_type",
            "in": "formData",
            "description": "ask user for the blood type of the patient",
            "enum": [
              "A",
              "B",
              "AB",
              "O"
            ],
            "required": true,
            "type": "string"
          },
          {
            "name": "birth_place",
            "in": "formData",
            "description": "ask user for the birth place of the patient",
            "required": true,
            "type": "string"
          },
          {
            "name": "date_of_birth",
            "in": "formData",
            "description": "ask user for the birthdate of the patient",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "get": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "Get all campaigns",
        "description": "show all the campaigns",
        "operationId": "getCampaign",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/campaign/{campaign_id}/": {
      "get": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "Get campaign details by campaign id",
        "description": "show details of the campaign based by campaign id",
        "operationId": "campaignDetails",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "insert the campaign id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "update campaign",
        "description": "update selected campaign based by the campaign id",
        "operationId": "updateCampaign",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "insert the campaign id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "You can update even if only one parameter",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateCampaign"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "delete campaign",
        "description": "delete the selected campaign based by campaign id",
        "operationId": "deleteCampaign",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "insert the selected campaign id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/campaign/verify/{campaign_id}/": {
      "put": {
        "tags": [
          "Campaign Collection"
        ],
        "summary": "verify campaign",
        "description": "verify selected campaign based by the campaign id (only superuser and admin could perform this action!)",
        "operationId": "verifyCampaign",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "insert the campaign id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "You can update even if only one parameter",
            "required": true,
            "schema": {
              "$ref": "#/definitions/verifyCampaign"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "updateCampaign": {
      "type": "object",
      "properties": {
        "image": {
          "type": "file"
        },
        "address": {
          "type": "string",
          "example": "the patient's address"
        },
        "title": {
          "type": "string",
          "example": "the campaign's title"
        },
        "patient_name": {
          "type": "string",
          "example": "the patient's name"
        },
        "description": {
          "type": "string",
          "example": "the campaign's description"
        },
        "total_fund": {
          "type": "number",
          "example": 1000000
        },
        "disease_category": {
          "type": "string",
          "enum": [
            "Asthma",
            "Diarrhea",
            "Diabetes Mellitus",
            "Dengue Fever",
            "Poor Nutrition",
            "Hypertension",
            "Coronary Heart Disease",
            "Malaria",
            "Stroke",
            "Tuberculosis",
            "Covid19",
            "Dental Health Problem"
          ]
        },
        "due_date": {
          "type": "date"
        },
        "gender": {
          "type": "string",
          "enum": [
            "Male",
            "Female"
          ]
        },
        "blood_type": {
          "type": "string",
          "enum": [
            "A",
            "B",
            "AB",
            "O"
          ]
        },
        "birth_place": {
          "type": "date"
        },
        "date_of_birth": {
          "type": "date"
        }
      }
    },
    "verifyCampaign": {
      "type": "object",
      "required": [
        "verify"
      ],
      "properties": {
        "verify": {
          "type": "boolean",
          "example": true
        }
      }
    }
  },
...
}